iommu/tegra-smmu: Unwrap tegra_smmu_group_get
The tegra_smmu_group_get was added to group devices in different
SWGROUPs but now it handles devices in same SWGROUP also, which
means that it allocates a group every time a new SWGROUP comes.

So it's not necessary to have a duplicated fallback routine any
more. This patch just unwraps the function to clean it up.

Bug 200647809

Change-Id: I1a1cab80b51af5994e0df896300fea87633d62cb
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2403788
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit