video: tegra: nvmap: Remove procrank debugfs
Making NvMap as OOT module results in walk_page_range and vm_normal_page
as undefined symbol during compilation, as these functions are not
exported by the kernel. The call to this function is made in the
function, which shows memory usage, is called via procrank debugfs. Add
a flag, which does not compile this code for kstable.

JIRA TMM-2505

Bug 2879962

Change-Id: I21a40a8f2f1ff805f3312c24b3cc823c02623b59
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2382405
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit