arm64: defconfig: Enable MODVERSIONS
CONFIG_MODVERSIONS is used by ABI Monitoring tool "abidiff":
https://android.googlesource.com/kernel/build/+/refs/heads/master/abi/README.md

This patch enables module versioning as an ABI consistency check.
A CRC value of the full prototype for an exported symbol will be
created. When a module is loaded/used, the CRC values contained
in the kernel will be compared with similar values in the module;
if they are not equal, the kernel will refuse to load the module.

It'll be useful to catch changes in data structure visibility. If
modules change opaque data structures, i.e. data structures that
are not part of the KMI (Kernel Module Interface), kernel modules
will break after future changes to the structure.

It also adds a runtime check to avoid loading a module that isn't
KMI compatible with the kernel. This will prevent a hard-to-debug
runtime issue or kernel crash that will show up in the future.

Meanwhile, "abidiff" has some current limitations in identifying
ABI differences in certain convoluted cases that MODVERSIONS can
catch.

Bug 200653251

Change-Id: I115248af6859f81961f72b40665ef3301c36f897
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2406715
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>