DOWNSTREAM: kernel: dma: Store the dma flag
While creating the coherent memory for the carveout, we pass the
DMA_MEMORY_NOMAP flag, because of which memremap is not performed on
passed physical address. But this flag is not getting stored in
dma_coherent_mem struct. While allocating from this coherent memory
this flag is checked, and as it is not stored, it results in doing
memset on some random V.A. and results in seg fault. Resolve this
failure by storing this flag.

Bug 200652203
Bug 200654035

Change-Id: I131a9292b42359ef5f9dc0c3e689f9bccdf206c1
Signed-off-by: Ketan Patil <ketanp@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2408482
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>