tegra_udrm: make it build on kernel version 5.9
When kernel version is >= 5.8,
* Drop use of the deprecated drmP.h header file.
* Use drm_dev_put() for drm_dev_unref().
* drm_event_vblank is placed in a new union event, so accordingly
  modify the code.

There seems to be discrepancy between the folder name which has kernel
source code and and the version of kernel mentioned in source code.
Folder name is kernel-5.9 but the version in source code is 5.8.

Note that tegra_udrm is not guaranteed to get build on intermediate
kernel versions between 4.14 and 5.8.

Bug 200656822

Change-Id: Ic0b811f769d9355b63add1244228fa7c0d390be0
Signed-off-by: Yogish Kulkarni <yogishk@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2411764
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>