Revert "[TEMP] ASoC: rt5659: use RT5659_JD_NULL for HD header"
This reverts commit 8849311a73378a59f7297fbcb434f9fadea23dba which
was a temporary change done to complete v5.9 codec integrations.
Now all the integrations have been done and it is time to remove
temporary change to align with upstream. In doing so all instances
of the macro are replaced.

Upstream kernel uses a different macro (RT5659_JD_HDA_HEADER) for
Intel 10-pin HD header support and the same is reverted back here.
All related platform DTs need to be updated and it is taken care
by other patches in the series. Previous kernels (v4.14 and v4.9)
use different macro and the same need to be updated as well since
DT is common.

Bug 200646528

Change-Id: I97af41b8f0baba176b818e4bf7c6d5baab367238
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2413247
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>