t19x: common: Cleanup IRQ handling for RT565x
As per DT binding documentation of rt5659 codec, "interrupts" and
"interrupt-parent" bindings need to be used to specify codec IRQ
output. Explicit parsing of GPIO entry and mapping it to IRQ is
not required in the driver.

This patch removes "gpios" entry from the common super-module file
and the new suggested binding is added in platform specific files
by other patches in the series. This is done because not all
platforms use the same GPIO for codec IRQ. Already Jakku and Mccoy
are overriding this entry in platform specific files.

Bug 200646528

Change-Id: Ieb6f1407c3094236fa44f672ed8ce2ef43378c25
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/common/+/2413253
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sameer Pujar <spujar@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit