t19x: galen: Cleanup RT565x DT binding
As per DT binding documentation of rt5659 codec, "interrupts" and
"interrupt-parent" bindings need to be used to specify codec IRQ
output. Explicit parsing of GPIO entry and mapping it to IRQ is
not required in the driver and thus "gpios" entry is replaced with
the above mentioned DT bindings.

New suggested binding is added for super-module as well. So far it
depended on common super-module file. But generally platforms use
different GPIO for the IRQ binding and already Jakku/Mccoy override
this entry in their platform specific files. The "gpios" entries
from common super-module file are removed by other patch in the
series.

Upstream kernel does not expose a kernel-include header file for
RT565x codec. In downstream such a file is defined and macros are
referenced from it to specify some of the DT binding values for
codec. Since RT565x codec is maintained by Realtek, we need not
keep additional files which are not found in upstream. Thus direct
values are specified for following properties.
 - realtek,jd-src
 - realtek,dmic1-data-pin
 - realtek,dmic2-data-pin

Bug 200646528

Change-Id: I7884c38eadd97298e8d9021d3940c586f01efc5f
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/stardust-dts/+/2413254
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sameer Pujar <spujar@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit