t19x: mccoy: Cleanup RT565x DT binding
As per DT binding documentation of rt5659 codec, "interrupts" and
"interrupt-parent" bindings need to be used to specify codec IRQ
output. Explicit parsing of GPIO entry and mapping it to IRQ is
not required in the driver and thus "gpios" entry is replaced with
the above mentioned DT bindings.

Upstream kernel does not expose a kernel-include header file for
RT565x codec. In downstream such a file is defined and macros are
referenced from it to specify some of the DT binding values for
codec. Since RT565x codec is maintained by Realtek, we need not
keep additional files which are not found in upstream. Thus direct
values are specified for following properties.
 - realtek,jd-src
 - realtek,dmic1-data-pin
 - realtek,dmic2-data-pin

Bug 200646528

Change-Id: I4389a00118b9b5d8ec9411c164859c52f670a0fb
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/mccoy-dts/+/2413255
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sameer Pujar <spujar@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit