dt-bindings: tegra: Remove rt5659.h header
The RT5659 codec driver is maintained by Realtek and upstream kernel
does not have 'rt5659.h' header defined in dt-bindings include path
of kernel. Earlier in downstream this was added to make macro usage
simpler and is currently used on few platforms. There is no need to
keep this additional header in downstream and since NVIDIA is not the
maintainer of this driver, the header is being removed to align with
upstream code. The header inclusion is removed by other platform DT
patches in the series.

Bug 200646528

Change-Id: Id21a948e715bf7bfb17b432c1fd04b24e06a6d49
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/tegra/+/2413846
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit