t19x: jakku: Cleanup IRQ handling for RT565x
As per DT binding documentation of rt5659 codec, "interrupts" and
"interrupt-parent" bindings need to be used to specify codec IRQ
output. Explicit parsing of GPIO entry and mapping it to IRQ is
not required in the driver and thus "gpios" entry is replaced with
the above mentioned DT bindings.

Bug 200646528

Change-Id: I55b49f793ffc73eabfcd22c5edb00ed42a2fb2e2
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/platform/t19x/jakku-dts/+/2414380
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Mohan Kumar D <mkumard@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sameer Pujar <spujar@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit