dmaengine: tegra210-adma: Wrong address used for interrupt register
The address used for global interrupt clear register on Tegra186 and later
is incorrect. The reason for this is, there are some differences between
Tegra210 and Tegra186 with respect to register space and current offset
is not working for the mentioned register.

On Tegra186 and later, ADMA channels are virtualized and aliased into 4 64K
pages (Currently SW uses page-1 by default). A separate page carries global
and configuration registers for ADMA. Few registers are reshuffled as part
of it and moved to page specific space. Thus offset of these registers are
relative to the channel base offset and it needs to be taken into account
while updating. In summary the register space looks like below:

 - Tegra210
      |
      |--> channel registers
      |          ...
      |--> global registers
      |

 - Tegra186 (and later)
      |
      |--> global registers
      |          ...
      |--> page-1 channel registers
      |          ...
      |--> page-1 global registers
      |
      |          ...
      |
      |--> page-4 channel registers
      |          ...
      |--> page-4 global registers
      |

The above mentioned interrupt register belongs to page global address space
and defined offset is not working. Fix this by using reference of channel
base address while updating this register.

Bug 200664413

Change-Id: I14a810db544cb8b9444144d0bdb527636422887e
Reported-by: Ryan Bedwell <rbedwell@nvidia.com>
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2424007
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Ryan Bedwell <rbedwell@nvidia.com>
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit