mmc: host: Fix sequence to parse host ocr
Host OCR is currently set to a default max value spanning the entire
VDD range.
If vmmc regulator node or mmc-ocr-mask is not populated in the dt as in
case of FPGA/VDK, invalid vdd is requested as part of ios setup
resulting in boot crash. This change ensures host OCR is set in the
tegra layer only if a corresponding ocr-mask node is defined in the
dtsi.

Bug 3147004
Bug 3146893

Change-Id: I692eb4f7ba0d4141a9e87cb57c0c4e08940928a8
Signed-off-by: anrao <anrao@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2425116
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>