dts: t23x: fpga: enable cbb error checking for SCE
CBB error checking was kept disabled for SCE on initial SNAP's.
SCE was not present in FPGA then and enabling it was giving
errors. Enabling it now as the boot is working without any
error.
Also, correcting a typo to change CBB node's start address from
'cbb-en@2300000' to 'cbb-en@1300000'

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Change-Id: Ie8fa183c1857efc280e721705f2679a345b15bac
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t23x/+/2434490
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit