soc/tegra: fuse: Add CONFIG to downstream code
This change is used to not build by default downstream fuse code,
as it might cause a build break due to lacking of dependency. It
will be helpful whenever we attach the nvidia-repo overlay to any
new kernel that won't by default have downstream intergration to
correctly build with fuse code.

However, any existing kernel using this code should add this new
CONFIG to their defconfig files so as to keep the consistency of
fuse function.

Note that it also changes fuse-burn's CONFIG_SOC_TEGRA_PMC that's
originally from K4.14, as the Kconfig in K4.9 is using obj-y for
fuse-burn. So it'd be better to unify it, and it would be safe to
do so, given that CONFIG_SOC_TEGRA_PMC was not enabled in K4.14's
defconfig but being selected by CONFIG_ARCH_TEGRA_210_SOC and by
other older SoCs that we probably no longer support on TOT.

Bug 200591811
Bug 200663781

Change-Id: I4729bb8be7c6ba6cb75bbfeb417cbed38fbeb7a6
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2443668
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>