gpu: nvgpu: Add abstraction layer for NVHost fence
In preparation for adding support for the upstream Host1x driver,
add a layer of abstraction for calling the NVHost fence APIs. This
provides a clean interface for calling the appropriate APIs for either
the downstream NVHost fence APIs or the upstream DMA fence and Host1x
APIs, depending on which is being used.

Change-Id: I055aee3fe5d3af83ebb11126c0a862e17b21f94c
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvgpu/+/2450244
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Sagar Kamble <skamble@nvidia.com>
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>