[TEMP] arm64: defconfig: Build ACONNECT driver into kernel
Device hang is observed on T186 during SC7 suspend/resume cycle and this
is related to ACONNECT device which fails to resume. The reason for this
seems to be related to ACONNECT device trying to resume before BPMP
does.

The issue seems to go away if ACONNECT driver is built as part of
kernel image itself. The exact cause for above issue needs to be
debugged. Meanwhile to unblock SC7 coverage in general this patch
can be used.

Bug 200664943
Bug 3173151

Change-Id: Ib5c7f068403205ebf37ae4a6ed5341c3537eca1e
Signed-off-by: Sameer Pujar <spujar@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2453116
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>