video: tegra: nvmap: cause sigsegv for priv anon mappings
Don't allow Private anonymous mappings as its not valid with
Nvmap. If the userspace creates private anonymous mapping and
then allocates Nvmap handle to share the buffer, it results in
system crash. Nvmap creates shared mappings by default which
is orthogonal to the Private anonymous mappings. During page
fault, cow handling is done due to the priv anon mapping from
userspace. But it results in crash as the shared flag is also
set which is not expected for such mapping.

Bug 3014767

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Change-Id: Ie61d4c23b0611fe79bd848226466d92922be07b8
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2457433
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit