dts: soc: t19x: Set wake83 IRQ to rising edge
IRQ binding of SW Wake (wake83) is updated to IRQ_TYPE_EDGE_RISING.
This is necessary for the interrupt to be raised at wake trigger and
subsequently the ISR to run in gpio-keys driver.

It should be noted that the interrupt gets raised only when waking up
from SC7. Asserting the wake (from SPE) during normal system operation
does not raise any interrupt.

Bug 200611152
Bug 200658583

Change-Id: I61a8d49e63bc9e9345f53f72f3546e1955144cd4
Signed-off-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t19x/+/2457968
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>