tegra: nvmap: fix wrong check of vmas in vma list
"atomic_add_unless" doesn't return older value of vmas
in the vma list, rather it returns true even if no of vmas in
vma list is > 1. Due to that during nvmap_vma_close(), nvmap handle
reference count is decremented by calling nvmap_handle_put()
in every call of nvmap_vma_close().

The expectation here is to use the function which returns older value
of vmas in the vma list so that when there is number of vmas in the vma
list is 1, refernce count should reach 0 and nvmap handle is freed.

"__atomic_add_unless" correctly returns older value of vmas
in vma list but its not defined in K5.10.
It's incorrectly ported to "atomic_add_unless". The equvivalent function
in K5.10 is "atomic_fetch_add_unless". Replacing
"atomic_add_unless" by "atomic_fetch_add_unless" in K5.10, the issue is
fixed.

Bug 200684122

Change-Id: I33dac48066ffad72d29a4090cd82e2a49151da29
Signed-off-by: Puneet Saxena <puneets@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2457994
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit