arm64: configs: disable CONFIG_RODATA_FULL_DEFAULT_ENABLED
rodata is debug feature in which read-only attributes applied to
VM areas of the linear alias of the backing pages. This prevents
code or read-only data being modified inadvertently. There is
Bug in the feature where sometime page attribute is not
updated as per map request causing crash in dma sync apis.
Disable the feature CONFIG RODATA_FULL_DEFAULT_ENABLED to avoid
the failure.

Bug 200677837

Change-Id: I0f713d60a48c48001c00dba1a7cb5ce50bcd2cd3
Signed-off-by: Pritesh Raithatha <praithatha@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2458214
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>