mmc: host: Parse SDMMC regulators in the probe
The current implementation of SDMMC driver defines a workqueue
in the probe for initializing the host controller.
However, if the regulator function call returns EPROBE_DEFER as
part of the host controller init, there is no way for the
SDMMC driver to be added in the deferred probe list.
Parse SDMMC regulators in the probe and add the driver
in the deferred probe list when regulator function calls
return EPROBE_DEFER.

Bug 200664943

Change-Id: I08d35d34f15dacc282eec506465a4a4f08aa91c1
Signed-off-by: Aniruddha Tvs Rao <anrao@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2458644
Tested-by: Bitan Biswas <bbiswas@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-by: Manish Bhardwaj <mbhardwaj@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>