dts: soc: t19x: Remove 'supports-clkreq' from SoC DT file
The device-tree property 'supports-clkreq' is to be updated based on the
connectivity of the CLKREQ signal from root port all the way upto the
endpoint. Though T194 based platforms have CLKREQ routing available upto
the slot, but there are PCIe cards that don't have support for CLKREQ.
So, it is very subjective as to when this DT entry should be added and
it pretty much depends on the endpoint being connected.

Bug 2561755

Change-Id: I26d72260f9a6808454542dcb55c460aab4bc8a29
Signed-off-by: Vidya Sagar <vidyas@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/soc/t19x/+/2464873
Reviewed-by: Manikanta Maddireddy <mmaddireddy@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>