platform: tegra: cbb: dont crash system for all cbb errors
This change is done to avoid crashing the system for errors which are
only notifications to indicate a transaction was not allowed to be
attempted. Such errors are:
1. SEC(Security violation) and Firewall error
2. DEC(Address decode)
3. DISC(Disconnected target or domain)
4. UNS(Unsupported request)
So, we can avoid crashing the system as the transaction didn't happen.

Bug 3191922

Signed-off-by: Sumit Gupta <sumitg@nvidia.com>
Change-Id: I6a6350aba8697176733a2fda2eb23b1af73f5c2d
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2459528
(cherry picked from commit 30fadeeb5c307c7f34ada98e564eb21c336bf360)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2466743
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit